Moved read_skill endpoints from main.py to dqm1 router
